refactor: adjust server wrappers
