+++ src/actions/actionUpdateUserHasOptedInToMembership.ts
@@ -9,7 +9,7 @@ import { prismaClient } from '@/utils/server/prismaClient'
 import { throwIfRateLimited } from '@/utils/server/ratelimit/throwIfRateLimited'
 import { getServerPeopleAnalytics } from '@/utils/server/serverAnalytics'
 import { parseLocalUserFromCookies } from '@/utils/server/serverLocalUser'
-import { withServerActionMiddleware } from '@/utils/server/withServerActionMiddleware'
+import { withServerActionMiddleware } from '@/utils/server/serverWrappers/withServerActionMiddleware'
 
 export const actionUpdateUserHasOptedInToMembership = withServerActionMiddleware(
   'actionUpdateUserHasOptedInToMembership',
